requirements.txt
requests
python-dateutil
openai
streamlit